+++ sea-orm-migration/tests/common/migration/m20220118_000001_create_cake_table.rs
@@ -38,8 +38,7 @@ impl MigrationTrait for Migration {
     }
 }
 
-/// Learn more at https://docs.rs/sea-query#iden
-#[derive(Iden)]
+#[derive(DeriveIden)]
 pub enum Cake {
     Table,
     Id,